Reorganize directroy structure
