+++ .gdb_history
@@ -1,0 +1,1 @@
+bt
