+++ src/Packet/PacketBuilder.hh
@@ -1,0 +1,3 @@
+#ifndef __PACKET_BUILDER__
+#define __PACKET_BUILDER__
+#endif
